# openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Cliftin API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/routines": {
      "get": {
        "operationId": "listRoutines",
        "summary": "List planned routines",
        "tags": [
          "routines"
        ],
        "description": "Returns non-deleted routines and supports q, programId, weekId, exerciseId, one explicit sort value, and opaque cursor pagination.",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque cursor returned by a previous page."
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "exerciseId",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "programId",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "example": "upper",
            "in": "query",
            "name": "q",
            "required": false,
            "description": "Case-insensitive substring match on routine name."
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "name"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-name"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "weekId"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-weekId"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "Single sort field; prefix with - for descending order."
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "weekId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "id": {
                            "minimum": 1,
                            "type": "integer",
                            "example": 100
                          },
                          "isDeleted": {
                            "type": "boolean"
                          },
                          "isNext": {
                            "type": "boolean"
                          },
                          "name": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "program": {
                            "additionalProperties": false,
                            "type": "object",
                            "properties": {
                              "id": {
                                "minimum": 1,
                                "type": "integer",
                                "example": 1
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ]
                          },
                          "week": {
                            "additionalProperties": false,
                            "type": "object",
                            "properties": {
                              "id": {
                                "minimum": 1,
                                "type": "integer",
                                "example": 10
                              },
                              "number": {
                                "minimum": 1,
                                "type": "integer",
                                "example": 1
                              }
                            },
                            "required": [
                              "id",
                              "number"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "isDeleted",
                          "isNext",
                          "name",
                          "program",
                          "week"
                        ]
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/routines/{routineId}": {
      "get": {
        "operationId": "getRoutine",
        "summary": "Get planned routine detail",
        "tags": [
          "routines"
        ],
        "description": "Returns a routine with its normalized planned exercises and sets. Direct lookup can return a soft-deleted routine with isDeleted true.",
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "lb"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "kg"
                  ]
                }
              ]
            },
            "example": "lb",
            "in": "query",
            "name": "unit",
            "required": false,
            "description": "Weight display unit. Defaults to lb where applicable."
          },
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "example": 100,
            "in": "path",
            "name": "routineId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "allOf": [
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "id": {
                          "examples": [
                            100
                          ],
                          "minimum": 1,
                          "type": "integer"
                        },
                        "isDeleted": {
                          "type": "boolean"
                        },
                        "isNext": {
                          "type": "boolean"
                        },
                        "name": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "program": {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "examples": [
                                1
                              ],
                              "minimum": 1,
                              "type": "integer"
                            },
                            "name": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        },
                        "week": {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "examples": [
                                10
                              ],
                              "minimum": 1,
                              "type": "integer"
                            },
                            "number": {
                              "examples": [
                                1
                              ],
                              "minimum": 1,
                              "type": "integer"
                            }
                          },
                          "required": [
                            "id",
                            "number"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "isDeleted",
                        "isNext",
                        "name",
                        "program",
                        "week"
                      ]
                    },
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "exercises": {
                          "type": "array",
                          "items": {
                            "additionalProperties": false,
                            "type": "object",
                            "properties": {
                              "exerciseId": {
                                "anyOf": [
                                  {
                                    "description": "Exercise definition primary key.",
                                    "minimum": 1,
                                    "type": "integer"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "id": {
                                "description": "Exercise configuration primary key.",
                                "minimum": 1,
                                "type": "integer"
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "sets": {
                                "type": "array",
                                "items": {
                                  "additionalProperties": false,
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "anyOf": [
                                        {
                                          "minimum": 1,
                                          "type": "integer"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "reps": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "rpe": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "timeSeconds": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "weight": {
                                      "additionalProperties": false,
                                      "type": "object",
                                      "properties": {
                                        "unit": {
                                          "description": "Weight display unit. Defaults to lb where applicable.",
                                          "examples": [
                                            "lb"
                                          ],
                                          "anyOf": [
                                            {
                                              "type": "string",
                                              "enum": [
                                                "lb"
                                              ]
                                            },
                                            {
                                              "type": "string",
                                              "enum": [
                                                "kg"
                                              ]
                                            }
                                          ]
                                        },
                                        "value": {
                                          "anyOf": [
                                            {
                                              "type": "number"
                                            },
                                            {
                                              "type": "null"
                                            }
                                          ]
                                        }
                                      },
                                      "required": [
                                        "unit",
                                        "value"
                                      ]
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "reps",
                                    "rpe",
                                    "timeSeconds",
                                    "weight"
                                  ]
                                }
                              }
                            },
                            "required": [
                              "exerciseId",
                              "id",
                              "name",
                              "sets"
                            ]
                          }
                        }
                      },
                      "required": [
                        "exercises"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": []
}